requirements.txt
absl-py==0.13.0
alembic==1.4.1
amqp==2.6.1
anyio==3.3.0
apache-airflow==2.1.3
apache-airflow-providers-cncf-kubernetes==2.0.1
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-http==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-sqlite==2.0.0
apispec==3.3.2
appdirs==1.4.4
argcomplete==1.12.3
astunparse==1.6.3
atpublic==2.3
attrs==20.3.0
Babel==2.9.1
billiard==3.6.4.0
blinker==1.4
cachetools==4.2.2
cattrs==1.5.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
clickclick==20.10.2
cloudpickle==1.6.0
colorama==0.4.4
colorlog==6.4.1
commonmark==0.9.1
configobj==5.0.6
croniter==1.0.15
cryptography==3.4.8
cycler==0.10.0
databricks-cli==0.15.0
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.3.4
diskcache==5.2.1
distro==1.6.0
dnspython==2.1.0
docker==5.0.0
docutils==0.16
dpath==2.0.1
dulwich==0.20.25
dvc==2.6.4
email-validator==1.1.3
entrypoints==0.3
Flask==1.1.4
Flask-AppBuilder==3.3.2
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flatbuffers==1.12
flatten-dict==0.4.2
flower==0.9.7
flufl.lock==3.2
fsspec==2021.8.1
ftfy==6.0.3
funcy==1.16
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grandalf==0.6
graphviz==0.17
grpcio==1.39.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
httpcore==0.13.6
httpx==0.19.0
humanize==3.11.0
idna==3.2
importlib-metadata==4.7.1
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.16
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==4.6.11
kubernetes==11.0.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
mailchecker==4.0.11
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.4.3
mlflow==1.20.1
mysqlclient==2.0.3
nanotime==0.5.2
networkx==2.6.2
numpy==1.19.5
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pathspec==0.8.1
pendulum==2.1.2
phonenumbers==8.12.31
Pillow==8.3.1
ply==3.11
prison==0.1.3
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.2
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
pygit2==1.6.1
Pygments==2.10.0
pygtrie==2.4.2
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-benedict==0.24.2
python-daemon==2.3.0
python-dateutil==2.8.2
python-editor==1.0.4
python-fsutil==0.5.0
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.7.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.1
setproctitle==1.2.2
shortuuid==1.0.1
shtab==1.4.1
six==1.15.0
sklearn==0.0
smmap==4.0.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.1
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
tqdm==4.62.2
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.6
vine==1.3.0
voluptuous==0.12.1
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xmltodict==0.12.0
zc.lockfile==2.0
zipp==3.5.0
